Remove walletAcountActivationPaymentInfo from redux
